+++ extensions/pi-gmail/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- `readOnly` setting to disable `send` and `send_draft` actions while preserving inbox and draft access
+
 ## [0.3.0] - 2026-05-08
 
 ### Changed